Part 3 – Queue Interface
The Ride class has been updated to manage the waiting queue using a queue, store the visitors in the queue, meet the FIFO queue requirements, and improve the methods of adding tourists to the queue, removing tourists, and printing the queue. In addition, the Ride and Visitor objects are created in partThree of the AssignmentTwo class and the methods are called.