Part 6 and Part 7 - Writing and Reading file
The content of these two parts is about import and export of files, so we will study them together. In Part 6, I created a method exportRideHistory to write tourists' information into a file. In Part 7, I read and output the content written in Part 6. There are various problems with file handling, so I've added appropriate exception handling and error messages to both methods. In addition, I created a new Ride and corresponding object in the main function, and called the import and export methods to prove the function.